do a disgusting amount of work for something simple tests are broken
