+++ big_scape/__main__.py
@@ -1,0 +1,7 @@
+"""Main file executed when the package is run as a module
+"""
+
+import big_scape.main
+
+if __name__ == "__main__":
+    big_scape.main.main()
