+++ big_scape/cli/cli_common_options.py
@@ -1,6 +1,7 @@
 """Click parameter that are shared between multiple BiG-SCAPE modules"""
 
 # from python
+import os
 import click
 from multiprocessing import cpu_count
 from pathlib import Path
@@ -18,7 +19,6 @@ from .cli_validations import (
     validate_classify,
     validate_output_dir,
 )
-from .constants import ROOT_DIR
 
 
 def common_all(fn):
@@ -36,9 +36,11 @@ def common_all(fn):
             type=click.Path(
                 exists=True, dir_okay=False, file_okay=True, path_type=Path
             ),
-            default="./config.yml",
+            default=Path(os.getcwd()) / "./config.yml",
             help="Path to BiG-SCAPE config file, which stores values for a "
-            "series of advanced use parameters. (default: [BiG-SCAPE directory]/config.ini).",
+            "series of advanced use parameters. (default: [working directory]/config.yml). "
+            "If this file does not exist, uses a default file packaged with BiG-SCAPE "
+            "([big_scape package]/config.yml).",
         ),
         # diagnostic parameters
         click.option(
