--- big_scape/cli/constants.py
@@ -1,3 +1,0 @@
-"""Contains constant values"""
-
-DB_SCHEMA_PATH = "big_scape/data/schema.sql"
