+++ big_scape/data/sqlite.py
@@ -24,7 +24,7 @@ import tqdm
 import click
 
 # from other modules
-from big_scape.cli.constants import DB_SCHEMA_PATH
+import big_scape.paths as bs_paths
 from big_scape.errors import DBClosedError, DBAlreadyOpenError
 
 
@@ -71,7 +71,7 @@ class DB:
         if not DB.connection:
             raise RuntimeError("DB.connection is None")
 
-        creation_queries = read_schema(Path(DB_SCHEMA_PATH))
+        creation_queries = read_schema(Path(bs_paths.DB_SCHEMA_FILE))
 
         for creation_query in creation_queries:
             DB.connection.execute(text(creation_query))
